% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLoops.R
\name{AssessOverallLoopSimilarity}
\alias{AssessOverallLoopSimilarity}
\title{Compare all loop sequences across multiple antibodies}
\usage{
AssessOverallLoopSimilarity(
  antibodies,
  wH1 = 1/6,
  wH2 = 1/6,
  wH3 = 1/6,
  wL1 = 1/6,
  wL2 = 1/6,
  wL3 = 1/6
)
}
\arguments{
\item{antibodies}{A list of antibodies}

\item{wH1}{The weighting for the H1 loop}

\item{wH2}{The weighting for the H2 loop}

\item{wH3}{The weighting for the H3 loop}

\item{wL1}{The weighting for the L1 loop}

\item{wL2}{The weighting for the L2 loop}

\item{wL3}{The weighting for the L3 loop}
}
\value{
A matrix antibodies x antibodies denoting the similarity of the loops
between each antibody pair
}
\description{
Assess the similarity between all loops of multiple antibodiesm, allowing for
loops to be weighted differently. Returns a matrix of antibodies x antibodies
denoting the similarity of the loops between each antibody pair.
}
\details{
Note, weights (wH1..wH3) must sum to 1
}
\examples{
# Read in 2 antibodies
path7x94 <- system.file("extdata", "7x94_imgt.pdb", package = "ReadAb")
antibody1 <- ReadAntibody(pdb = path7x94,
                          numbering = "IMGT",
                          heavy = "H",
                          light = "L",
                          antigen = "A")

path7x96 <- system.file("extdata", "7x96_imgt.pdb", package = "ReadAb")
antibody2 <- ReadAntibody(pdb = path7x96,
                          numbering = "IMGT",
                          heavy = "H",
                          light = "L",
                          antigen = "A")

# Compare the loops of antibody1 and antibody2, weighting H3 higher than
# other loops
AssessOverallLoopSimilarity(list(Antibody1 = antibody1, Antibody2 = antibody2),
                            wH1 = 0.1,
                            wH2 = 0.1,
                            wH3 = 0.5,
                            wL1 = 0.1,
                            wL2 = 0.1,
                            wL3 = 0.1)

}
