% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{DisplaySimilarityPlot}
\alias{DisplaySimilarityPlot}
\title{Display a heatmap of a matrix denoting the similarity between antibody loops}
\usage{
DisplaySimilarityPlot(similarity_matrix, loop)
}
\arguments{
\item{similarity_matrix}{A similarity matrix produced by
\code{assessLoopSimilarity} or \code{assessOverallLoopSimilarity}}

\item{loop}{The name of the loop being compared (should be one of 'H1', 'H2',
'H3', 'L1', 'L2', 'L3', or 'all')}
}
\description{
Display a heatmap of a matrix denoting the similarity between antibody loops
}
\examples{
# Read in 2 antibodies
antibody1 <- ReadAntibody(pdb = "data/7uja_chothia.pdb", 
                          numbering = "Chothia",
                          heavy = c("B", "E", "G", "I", "L", "N"),
                          light = c("D", "F", "H", "J", "M", "O"),
                          antigen = c("A", "K", "C"))

antibody2 <- ReadAntibody(pdb = "data/1ahw_chothia.pdb",
                          numbering = "Chothia",
                          heavy = c("B", "E"),
                          light = c("A", "D"),
                          antigen = c("C", "F"))
                          
# Example 1:
# Compare the H3 loops of antibody1 and antibody2 and display the plot
H3Sim <- assessLoopSimilarity(list(Antibody_1 = antibody1, 
                                   Antibody2 = antibody2), 
                              'H3')

DisplaySimilarityPlot(H3Sim, 'H3')

# Example 2:
# Compare the loops of antibody1 and antibody2, weighting H3 higher than 
# other loops
overallSim <- assessOverallLoopSimilarity(list(Antibody_1 = antibody1, Antibody2 = antibody2),
                                          wH1 = 0.1,
                                          wH2 = 0.1,
                                          wH3 = 0.5,
                                          wL1 = 0.1,
                                          wL2 = 0.1,
                                          wL3 = 0.1)

DisplaySimilarityPlot(overallSim, 'all')  

}
