% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAntibody.R
\name{.GetLoopAtoms}
\alias{.GetLoopAtoms}
\title{Get atoms in a loop}
\usage{
.GetLoopAtoms(loopName, pdb, chainTypes, ranges)
}
\arguments{
\item{loopName}{The name of the loop for which to select the atoms (ex. 'H1')}

\item{pdb}{An open bio3d pdb object}

\item{chainTypes}{A data frame with 2 columns (H and L) that stores the
identifiers of the heavy (H) and light (L) chains in the pdb}

\item{ranges}{A list of residue index ranges for each CDR loop}
}
\value{
A list of class "select" from bio3d of atoms in the pdb that are part
of the loop named loopName
}
\description{
Select the atoms in the loops named loopName from the pdb. The types of each
chain (i.e. H (heavy) vs L (light)) are denoted in chainTypes, and the
residue index ranges for each CDR according to the selected numbering scheme
are stored in ranges.
}
\examples{
# Read in a PDB file and note the chain types in a data frame
pdb_path <- system.file("extdata", "1ahw_chothia.pdb", package = ReadAb)
pdb <- bio3d::read.pdb(pdbPath)
chainTypes <- data.frame(H = c('B', 'E'), L = c('A', 'D'))

# Get the CDR index ranges for the numbering scheme used in the PDB
ranges <- GetNumberingRange('Chothia')

# Get the atoms in H1 loops in the pdb
H1Atoms <- .GetLoopAtoms('H1', pdb, chainTypes, ranges)

}
