% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibodyClass.R
\name{getLoopSequence}
\alias{getLoopSequence}
\title{Get the amino acid sequence of a loop from an antibody}
\usage{
getLoopSequence(antibody, loop)
}
\arguments{
\item{antibody}{A list of class 'antibody'}

\item{loop}{The name of a loop (one of 'H1'...'L3')}
}
\value{
The sequence of the given loop in the antibody as a string
}
\description{
Get the amino acid sequence the given loop in the given antibody.
If there are more than one copies of the loop present, the sequence of the
earliest one in the PDB will be returned.
}
\examples{
antibody <- ReadAntibody(pdb = "data/7uja_chothia.pdb", 
                         numbering = "Chothia",
                         heavy = c("B", "E", "G", "I", "L", "N"),
                         light = c("D", "F", "H", "J", "M", "O"),
                         antigen = c("A", "K", "C"))

# Get the sequence of the H1 loop in antibody in the first chain that appears
# in the PDB (in this case, "B")
getLoopSequence(antibody = antibody, loop = 'H1')

}
