% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAntibody.R
\name{GetNumberingRange}
\alias{GetNumberingRange}
\title{Get the index ranges for all CDRs for the specified numbering scheme
Note: AHo and Honneger schemes are equivalent}
\usage{
GetNumberingRange(scheme)
}
\arguments{
\item{scheme}{valid numbering scheme (one of 'Chothia', 'AHo', 'Honneger', 'Kabat'
'Martin', or 'IMGT')}
}
\value{
A list of residue index ranges for each CDR according to the numbering
scheme. Formatted as such (s & e denote start and end):
list(L1 = c(L1_s, L1_e), L2 = c(L2_s, L2_e), L3 = c(L3_s, L3_e),
H1 = c(H1_s, H1_e), H2 = c(H2, H2_e), H3 = c(H3_s, H3_e))
}
\description{
Get the index ranges for all CDRs for the specified numbering scheme
Note: AHo and Honneger schemes are equivalent
}
\examples{
GetNumberingRange('Chothia')

}
