% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibodyClass.R
\name{IsValidColor}
\alias{IsValidColor}
\title{Determines validity of a color}
\usage{
IsValidColor(color)
}
\arguments{
\item{color}{a color input by the user which must be checked for validity}
}
\value{
TRUE iff the color is valid in R. FALSE otherwise.
}
\description{
Determines whether the given color is valid in R
}
\examples{
Example 1:
IsValidColor('red') # This should return TRUE

Example 2:
IsValidColor('not a color') # This should return FALSE

}
