% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAntibody.R
\name{ReadAntibody}
\alias{ReadAntibody}
\title{Read in an antibody from a PDB}
\usage{
ReadAntibody(
  pdbPath,
  numbering = "Chothia",
  heavy = "H",
  light = "L",
  antigen = NULL
)
}
\arguments{
\item{pdbPath}{A string indicating the path to the antibody PDB file}

\item{numbering}{A string indicating the renumbering scheme applied to the
antibody at pdbPath. Default value is 'Chothia'}

\item{heavy}{The identifiers of the heavy chains (may be a vector or single
value)}

\item{light}{The identifiers of the heavy chains (may be a vector or single
value)}

\item{antigen}{The identifiers of the antigen chains (may be a vector, single
value, or NULL to indicate no antigen is present)}
}
\value{
A list of class 'antibody' with the following components:
- pdb: A list of class "pdb" from bio3d of the entire pdb
- loops: A list of six components:
- H1: A list of class "pdb" from bio3d of all H1 loops from the pdb
- H2: ... all H2 loops from the pdb
- H3: ... all H3 loops from the pdb
- L1: ... all L1 loops from the pdb
- L2: ... all L2 loops from the pdb
- L3: ... all L3 loops from the pdb
- antigen_chains: The names of all antigen chains in the pdb (or
NULL if none are specified)
- colors: A list mapping each component of the pdb (all loops,
antigen, and other) to a color for visualization
}
\description{
A function that reads in an antibody or antibody-antigen complex from a PDB,
given the path and re-numbeirng scheme (Kabat, Chothia, IMGT, Martin, AHo,
Honneger) and saves it in as an Antibody object with CDR loops and antigen
identified.
}
\details{
Note AHo and Honneger are equivalent, but both names are often used, so both
are supported
}
\examples{
chothiaPath <- system.file("extdata", "7uja_chothia.pdb", package = "ReadAb")
chothiaAntibody <- ReadAntibody(pdb = chothiaPath,
                                 numbering = "Chothia",
                                 heavy = c("B", "E", "G", "I", "L", "N"),
                                 light = c("D", "F", "H", "J", "M", "O"),
                                 antigen = c("A", "K", "C"))

imgtPath <- system.file("extdata", "7ru4_imgt.pdb", package = "ReadAb")
imgtAntibody <- ReadAntibody(pdb = imgtPath,
                              numbering = "IMGT",
                              heavy = "H",
                              light = "L",
                              antigen = "A")

# Print the read antibodies
print("Chothia antibody:")
print(chothiaAntibody)

print("IMGT antibody: ")
print(imgtAntibody)


}
