% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runReadAb.R
\name{RunReadAb}
\alias{RunReadAb}
\title{Launch Shiny App for ReadAb}
\usage{
RunReadAb()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for ReadAb
This app allows the user to upload antibody structures in PDB files and
analyze the similarity of thier loop sequences.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

ReadAb::RunReadAb()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
