% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{VisualizeAntibody}
\alias{VisualizeAntibody}
\title{Visualize the provided antibody in 3D}
\usage{
VisualizeAntibody(antibody, mode = "all_atoms")
}
\arguments{
\item{antibody}{A list of class 'antibody'}

\item{mode}{A string descibing the types of atoms to display. Must be one of:
"all_atoms": to display all atoms
"heavy": to display all heavy (non-hydrogen) atoms
"backbone": to omit sidechain and hydrogen atoms from the plot}
}
\description{
Displays the atoms of the antibody in 3D with atoms color coded using the
colors element of the antibody. Can display in 3 modes: "all_atoms" to show
all atoms of the antibody, "heavy" to show only heavy atoms (all atoms other
than hydrogen), or "backbone" to omit sidechain and hydrogen atoms.
}
\examples{
pdbPath <- system.file("extdata", "7uja_chothia.pdb", package = "ReadAb")
antibody <- ReadAntibody(pdb = pdbPath,
                         numbering = "Chothia",
                         heavy = c("B", "E", "G", "I", "L", "N"),
                         light = c("D", "F", "H", "J", "M", "O"),
                         antigen = c("A", "K", "C"))

# View all atoms of the antibody
VisualizeAntibody(antibody, mode = "all_atoms")

# View the antibody's heavy atoms
VisualizeAntibody(antibody, mode = "heavy")

# View the backbone only
VisualizeAntibody(antibody, mode = "backbone")

TODO add bonds

}
