% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLoops.R
\name{AlignLoop}
\alias{AlignLoop}
\title{Align the sequences of one loop of multiple antibodies}
\usage{
AlignLoop(antibodies, loop)
}
\arguments{
\item{antibodies}{A list of lists of class antibody (i.e. a list of antibodies)}

\item{loop}{The name of the loop to align (one of 'H1', 'H2', 'H3', 'L1',
'L2', 'L3')}
}
\value{
An AAStringSet of the aligned loops
}
\description{
Takes multiple antibodies and the name of a loop (one of 'H1'...'L3') and
returns an alignment of the sequences of this loop in the antibodies
}
\examples{
# Read in two antibodies
path7x94 <- system.file("extdata", "7x94_imgt.pdb", package = ReadAb)
antibody1 <- ReadAntibody(pdb = path7x94,
                          numbering = "IMGT",
                          heavy = "H",
                          light = "L",
                          antigen = "A")

path7x96 <- system.file("extdata", "7x96_imgt.pdb", package = ReadAb)
antibody1 <- ReadAntibody(pdb = path7x96,
                          numbering = "IMGT",
                          heavy = "H",
                          light = "L",
                          antigen = "A")

# Align their H1 loops
AlignLoop(list(antibody1, antibody2), "H1")

}
