% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLoops.R
\name{alignLoop}
\alias{alignLoop}
\title{Align the sequences of one loop of multiple antibodies}
\usage{
alignLoop(antibodies, loop)
}
\arguments{
\item{antibodies}{A list of lists of class antibody (i.e. a list of antibodies)}

\item{loop}{The name of the loop to align (one of 'H1', 'H2', 'H3', 'L1',
'L2', 'L3')}
}
\value{
An AAStringSet of the aligned loops
}
\description{
Takes multiple antibodies and the name of a loop (one of 'H1'...'L3') and
returns an alignment of the sequences of this loop in the antibodies
}
\examples{
# Read in two antibodies
antibody1 <- ReadAntibody(pdb = "data/7uja_chothia.pdb", 
                          numbering = "Chothia",
                          heavy = c("B", "E", "G", "I", "L", "N"),
                          light = c("D", "F", "H", "J", "M", "O"),
                          antigen = c("A", "K", "C"))

antibody2 <- ReadAntibody(pdb = "data/1ahw_chothia.pdb",
                          numbering = "Chothia",
                          heavy = c("B", "E"),
                          light = c("A", "D"),
                          antigen = c("C", "F"))

# Align their H1 loops
alignLoop(list(antibody1, antibody2), "H1")

}
