% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLoops.R
\name{assessLoopSimilarity}
\alias{assessLoopSimilarity}
\title{Assess the similarity between one loop across multiple antibodies. Returns
a matrix of antibodies x antibodies denoting the similarity of the loop
between each antibody pair.}
\usage{
assessLoopSimilarity(antibodies, loop)
}
\arguments{
\item{antibodies}{A list of antibodies}

\item{loop}{The name of a loop to compare across antibodies (must be one of 'H1', 'H2', 'H3', 'L1', 'L2', or 'L3')}
}
\value{
A matrix antibodies x antibodies denoting the similarity of the loop
between each antibody pair
}
\description{
Assess the similarity between one loop across multiple antibodies. Returns
a matrix of antibodies x antibodies denoting the similarity of the loop
between each antibody pair.
}
\examples{
# Read in 2 antibodies
antibody1 <- ReadAntibody(pdb = "data/7uja_chothia.pdb", 
                          numbering = "Chothia",
                          heavy = c("B", "E", "G", "I", "L", "N"),
                          light = c("D", "F", "H", "J", "M", "O"),
                          antigen = c("A", "K", "C"))

antibody2 <- ReadAntibody(pdb = "data/1ahw_chothia.pdb",
                          numbering = "Chothia",
                          heavy = c("B", "E"),
                          light = c("A", "D"),
                          antigen = c("C", "F"))
                          
# Compare the H3 loops of antibody1 and antibody2
assessLoopSimilarity(list(Antibody_1 = antibody1, Antibody2 = antibody2), 'H3')

}
