% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLoops.R
\name{AssessLoopSimilarity}
\alias{AssessLoopSimilarity}
\title{Compare a loop sequence across multiple antibodies}
\usage{
AssessLoopSimilarity(antibodies, loop)
}
\arguments{
\item{antibodies}{A list of antibodies}

\item{loop}{The name of a loop to compare across antibodies (must be one of 'H1', 'H2', 'H3', 'L1', 'L2', or 'L3')}
}
\value{
A matrix antibodies x antibodies denoting the similarity of the loop
between each antibody pair
}
\description{
Assess the similarity between one loop across multiple antibodies. Returns
a matrix of antibodies x antibodies denoting the similarity of the loop
between each antibody pair.
}
\examples{
# Read in 2 antibodies
path7x94 <- system.file("extdata", "7x94_imgt.pdb", package = "ReadAb")
antibody1 <- ReadAntibody(pdb = path7x94,
                          numbering = "IMGT",
                          heavy = "H",
                          light = "L",
                          antigen = "A")

path7x96 <- system.file("extdata", "7x96_imgt.pdb", package = "ReadAb")
antibody1 <- ReadAntibody(pdb = path7x96,
                          numbering = "IMGT",
                          heavy = "H",
                          light = "L",
                          antigen = "A")

# Compare the H3 loops of antibody1 and antibody2
AssessLoopSimilarity(list(Antibody_1 = antibody1, Antibody2 = antibody2), 'H3')

}
