% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{DisplaySimilarityPlot}
\alias{DisplaySimilarityPlot}
\title{Display heatmap of loop similarity matrix}
\usage{
DisplaySimilarityPlot(similarityMatrix, loop)
}
\arguments{
\item{similarityMatrix}{A similarity matrix produced by
\code{assessLoopSimilarity} or \code{assessOverallLoopSimilarity}}

\item{loop}{The name of the loop being compared (should be one of 'H1', 'H2',
'H3', 'L1', 'L2', 'L3', or 'all')}
}
\description{
Display a heatmap of a matrix denoting the similarity between antibody loops
}
\examples{
# Read in 3 antibodies
path7x94 <- system.file("extdata", "7x94_imgt.pdb", package = "ReadAb")
antibody1 <- ReadAntibody(pdb = path7x94,
                          numbering = "IMGT",
                          heavy = "H",
                          light = "L",
                          antigen = "A")

path7x96 <- system.file("extdata", "7x96_imgt.pdb", package = "ReadAb")
antibody1 <- ReadAntibody(pdb = path7x96,
                          numbering = "IMGT",
                          heavy = "H",
                          light = "L",
                          antigen = "A")

path8sau <- system.file("extdata", "8sau_chothia.pdb", package = "ReadAb")
antibody3 <- ReadAntibody(pdb = path8sau,
                          numbering = "Chothia",
                          heavy = c("C", "H", "M"),
                          light = c("D", "I", "N"),
                          antigen = c("A", "F", "K"))

# Example 1:
# Compare the H3 loops of the antibodies and display the plot
H3Sim <- assessLoopSimilarity(list(Antibody1 = antibody1,
                                   Antibody2 = antibody2,
                                   Antibody3 = antibody3),
                              'H3')

DisplaySimilarityPlot(H3Sim, 'H3')

# Example 2:
# Compare the loops of antibody1 and antibody2, weighting H3 higher than
# other loops
overallSim <- assessOverallLoopSimilarity(list(Antibody1 = antibody1,
                                               Antibody2 = antibody2,
                                               Antibody3 = antibody3),
                                          wH1 = 0.1,
                                          wH2 = 0.1,
                                          wH3 = 0.5,
                                          wL1 = 0.1,
                                          wL2 = 0.1,
                                          wL3 = 0.1)

DisplaySimilarityPlot(overallSim, 'all')

}
