% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibodyHelpers.R
\name{GetLoopSequence}
\alias{GetLoopSequence}
\title{Get the amino acid sequence of a loop from an antibody}
\usage{
GetLoopSequence(antibody, loop, chain = NULL)
}
\arguments{
\item{antibody}{A list of class 'antibody'}

\item{loop}{The name of a loop (one of 'H1'...'L3')}

\item{chain}{The identifier of the chain to get the loop from. Default is
NULL, in which case the sequence of the earliest one in the PDB
will be returned.}
}
\value{
The sequence of the given loop in the antibody as a string
}
\description{
Get the amino acid sequence the given loop in the given antibody.
}
\examples{
pdbPath <- system.file("extdata", "7uja_chothia.pdb", package = "ReadAb")
antibody <- ReadAntibody(pdb = pdbPath,
                         numbering = "Chothia",
                         heavy = c("B", "E", "G", "I", "L", "N"),
                         light = c("D", "F", "H", "J", "M", "O"),
                         antigen = c("A", "K", "C"))

# Get the sequence of the H1 loop in antibody in the first chain that appears
# in the PDB (in this case, "B")
getLoopSequence(antibody = antibody, loop = 'H1')

}
